Time: 4 ms (42.62%) | Memory: 68.9 MB (98.09%) - LeetSync

+++ 643-maximum-average-subarray-i/maximum-average-subarray-i.java
@@ -6,7 +6,7 @@ class Solution {
         {
             sum+=nums[i];
         }
-        int max=sum;
+        double max=sum;
 
         for(int i=k;i<n;i++)
         {
@@ -14,7 +14,7 @@ class Solution {
 
             max=Math.max(max,sum);
         }
-
-        return (double) max/k;
+        
+        return max/k;
     }
 }
